Move PacketBuilder to a separated module
