+++ d2bs/kolbot/libs/modules/PacketBuilder.js
@@ -1,0 +1,66 @@
+/**
+ * @author Nishimura-Katsuo
+ * @description PacketBuilder, simpler way of sending a packet
+ */
+
+/*
+
+new PacketBuilder() - create new packet object
+
+Example (Spoof 'reassign player' packet to client):
+	new PacketBuilder().byte(0x15).byte(0).dword(me.gid).word(x).word(y).byte(1).get();
+
+Example (Spoof 'player move' packet to server):
+	new PacketBuilder().byte(0x3).word(x).word(y).send();
+*/
+/** @constructor */
+module.exports = function PacketBuilder() {
+	/* globals DataView ArrayBuffer */
+	if (this.__proto__.constructor !== PacketBuilder) {
+		throw new Error("PacketBuilder must be called with 'new' operator!");
+	}
+
+	let queue = [], count = 0;
+
+	let enqueue = (type, size) => (...args) => { // accepts any number of arguments
+		args.forEach(arg => {
+			if (type === 'String') {
+				arg = stringToEUC(arg);
+				size = arg.length + 1;
+			}
+
+			queue.push({type: type, size: size, data: arg});
+			count += size;
+		});
+
+		return this;
+	};
+
+	this.float = enqueue("Float32", 4);
+	this.dword = enqueue("Uint32", 4);
+	this.word = enqueue("Uint16", 2);
+	this.byte = enqueue("Uint8", 1);
+	this.string = enqueue("String");
+
+	this.buildDataView = () => {
+		let dv = new DataView(new ArrayBuffer(count)), i = 0;
+		queue.forEach(field => {
+			if (field.type === "String") {
+				for (let l = 0; l < field.data.length; l++) {
+					dv.setUint8(i++, field.data.charCodeAt(l), true);
+				}
+
+				i += field.size - field.data.length; // fix index for field.size !== field.data.length
+			} else {
+				dv['set' + field.type](i, field.data, true);
+				i += field.size;
+			}
+		});
+
+		return dv;
+	};
+
+	this.send = () => (sendPacket(this.buildDataView().buffer), this);
+	this.spoof = () => (getPacket(this.buildDataView().buffer), this);
+	this.get = this.spoof; // same thing but spoof has clearer intent than get
+};
